Add support for lit.exe lit.exe is the library generator for WiX. It works similar to light.exe but has fewer options. Factor the common bits into a base class, then inherent this in two separate tasks to deal with lit and light separately.
